User Profile Updated to use the new DB

+++ JtProject/src/main/java/ca/concordia/eats/controller/MainController.java
@@ -14,8 +14,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import javax.servlet.http.HttpSession;
 import java.util.List;
 
-import javax.servlet.http.HttpSession;
-
 @Controller
 public class MainController {
 
@@ -24,6 +22,7 @@ public class MainController {
 
     @GetMapping("/product/make-favorite")
     public String makeFavorite(@RequestParam("productid") int productId, @RequestParam("src") String sourcePage, HttpSession session) {
+        if (session.getAttribute("user") == null) return "userLogin";
         Customer customer = (Customer) session.getAttribute("user");
         productService.makeFavorite(customer.getUserId(), productId);
         customer.setFavorite(new Favorite(productService.fetchCustomerFavoriteProducts(customer.getUserId())));
@@ -32,6 +31,7 @@ public class MainController {
 
     @GetMapping("/product/remove-favorite")
     public String removeFavorite(@RequestParam("productid") int productId, @RequestParam("src") String sourcePage, HttpSession session) {
+        if (session.getAttribute("user") == null) return "userLogin";
         Customer customer = (Customer) session.getAttribute("user");
         productService.removeFavorite(customer.getUserId(), productId);
         customer.setFavorite(new Favorite(productService.fetchCustomerFavoriteProducts(customer.getUserId())));
@@ -40,13 +40,15 @@ public class MainController {
 
     @GetMapping("/favorites")
     public String fetchCustomerFavoriteProducts(HttpSession session, Model model) {
+        if (session.getAttribute("user") == null) return "userLogin";
         Customer customer = (Customer) session.getAttribute("user");
         model.addAttribute("favoriteProducts", customer.getFavorite().getCustomerFavoritedProducts());
         return "/favorites";
     }
 
     @GetMapping("/search")
-    public String search(@RequestParam("query") String query, Model model, HttpSession session ) {
+    public String search(@RequestParam("query") String query, Model model, HttpSession session) {
+        if (session.getAttribute("user") == null) return "userLogin";
         User user = (User) session.getAttribute("user");
         List<Product> products = productService.search(query, user.getUserId());
         model.addAttribute("products", products);
